pyproject.toml
[tool.poetry]
name = "startup-evaluator"
version = "0.1.0"
description = ""
authors = ["Dmitry <erofeevdma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
pydantic = "^2.9.2"
openai = "^1.52.0"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
aiogram = "^3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
